project/topk-music/requirements.txt
Flask
flask-restful
redis
cassandra-driver
